Changed build of script files.
Got rid of %VAR% variables again. Instead, the currently set CMake variables in the CMakeLists.txt file are dumped to a CMake script file which is included at the top of the build script. This avoids the necessity of re-configuring the entire build system whenever a .in script file was modified. Further, there is no need to distinguish between variables which are replaced during the configure step, and ones which are replaced during the build step.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1485 3c61e6fb-ae43-4039-8cf7-c895b3257837