Fixed test of MATLAB CMake tools to add own python module search path at the beginning of PYTHONPATH.
An older installation of BASIS which does not have the proper __init__.py file in lib/python/sbia/basis/ probably prevents Python from finding the testaux module used by the test, which is in the Testing/lib/python/sbia/basis/ directory. Before, we appended this path to PYTHONPATH, so it should be inserted at the begin of the search path.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1546 3c61e6fb-ae43-4039-8cf7-c895b3257837