Fixed copying of API documentation files in CTest script.
Instead of using the file(INSTALL) command, use cmake -E to copy the API documentation files. This allows the setting of the DESTDIR environment variable which will then only influence the installation of the project files, not the installation of the API documentation. For the latter, the actual path is given as argument to the doxygen option of the CTest script or the default of SBIA is used.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1672 3c61e6fb-ae43-4039-8cf7-c895b3257837