Fixed bugs in subprocess module.
Including a work-around for a bug regarding fork() in the libgcov library on Mac OS X 10.6 and a fix of the problem with TerminateProcess() on Windows XP.

Further, disabled the build configurations Coverage and MemCheck on Windows. Yet need to find a way how to do coverage analysis, respectively, memory checks on Windows.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1684 3c61e6fb-ae43-4039-8cf7-c895b3257837