Only optionally auto-prefix public headers.
Due to some unwanted side-effects when all public header files are copied to the build tree (potential to edit wrong file after compiler error, files in build tree not considered for report of uncovered files,…), by default, only header files with the .in suffix are configured and written to the build tree. All other header files are used from the source tree. The option BASIS_AUTO_PREFIX_INCLUDES can, however, be set to TRUE to use the previous implementation. A problem with using the wrong cached CMake variables in modules was fixed.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1727 3c61e6fb-ae43-4039-8cf7-c895b3257837