Add changelog generation target to end of basis_project_impl() which is added regardless of BUILD_DOCUMENTATION option.
The changelog target is treated separate from the other documentation because its generation requires the presence of the Subversion history (or Git history once Git is supported by BASIS), may required the user to enter their credentials, and further is timely expensive due to the slowness of the svn log command. Therefore, even when BUILD_DOCUMENTATION is OFF, add the changelog target such that the change log file can be generated manually by building this target.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@1790 3c61e6fb-ae43-4039-8cf7-c895b3257837