Refactor basis_add_script*() functions. Rename BASIS_SOURCE variable to BASIS_BASH_UTILITIES.
The BASIS_BASH_UTILITIES variable may be set as environment variable as part of the BASIS installation. The BASH scripts that make use of the BASIS utilities for BASH shall then be generally functional. When build by BASIS, the environment variable is overwritten in each script to include the proper basis.sh module instead.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@2396 3c61e6fb-ae43-4039-8cf7-c895b3257837