Revise BASIS utilities. Intermediate stage.
The BASIS utilities have been completely revised, including an API change of some of the functions. Next step is to combine the different project-specific templates into a single template in order to reduce the number of source files that are "magically" added by BASIS to a project to a minimum.

git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/BASIS/trunk@2399 3c61e6fb-ae43-4039-8cf7-c895b3257837